dodanie README

+++ src/Controller/FlashcardController.php
@@ -10,15 +10,39 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
 
 #[Route('/flashcard')]
+#[IsGranted('ROLE_USER')]
 final class FlashcardController extends AbstractController
 {
     #[Route(name: 'app_flashcard_index', methods: ['GET'])]
-    public function index(FlashcardRepository $flashcardRepository): Response
+    public function index(Request $request, FlashcardRepository $flashcardRepository, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+        $categoryId = $request->query->get('category');
+        
+        $categories = $entityManager->getRepository(\App\Entity\Category::class)->findAll();
+        
+        if ($categoryId && $categoryId !== 'all') {
+            $flashcards = $flashcardRepository->findBy([
+                'user' => $user,
+                'category' => $categoryId
+            ]);
+            $selectedCategory = $entityManager->getRepository(\App\Entity\Category::class)->find($categoryId);
+            $selectedCategoryName = $selectedCategory ? $selectedCategory->getName() : 'All';
+        } else {
+            $flashcards = $flashcardRepository->findBy(['user' => $user]);
+            $selectedCategoryName = 'All';
+            $categoryId = 'all';
+        }
+        
         return $this->render('flashcard/index.html.twig', [
-            'flashcards' => $flashcardRepository->findAll(),
+            'flashcards' => $flashcards,
+            'categories' => $categories,
+            'selectedCategory' => $categoryId,
+            'selectedCategoryName' => $selectedCategoryName,
         ]);
     }
 
@@ -30,6 +54,7 @@ final class FlashcardController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $flashcard->setUser($this->getUser());
             $entityManager->persist($flashcard);
             $entityManager->flush();
 
@@ -45,6 +70,10 @@ final class FlashcardController extends AbstractController
     #[Route('/{id}', name: 'app_flashcard_show', methods: ['GET'])]
     public function show(Flashcard $flashcard): Response
     {
+        if ($flashcard->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You can only view your own flashcards.');
+        }
+
         return $this->render('flashcard/show.html.twig', [
             'flashcard' => $flashcard,
         ]);
@@ -53,6 +82,10 @@ final class FlashcardController extends AbstractController
     #[Route('/{id}/edit', name: 'app_flashcard_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Flashcard $flashcard, EntityManagerInterface $entityManager): Response
     {
+        if ($flashcard->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You can only edit your own flashcards.');
+        }
+
         $form = $this->createForm(FlashcardForm::class, $flashcard);
         $form->handleRequest($request);
 
@@ -71,6 +104,10 @@ final class FlashcardController extends AbstractController
     #[Route('/{id}', name: 'app_flashcard_delete', methods: ['POST'])]
     public function delete(Request $request, Flashcard $flashcard, EntityManagerInterface $entityManager): Response
     {
+        if ($flashcard->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You can only delete your own flashcards.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$flashcard->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($flashcard);
             $entityManager->flush();
